requirements.txt
fastapi
uvicorn
streamlit
pydantic
google-genai
python-dotenv
requests
